Moving compiler check to setup.py
