+++ setup.py
@@ -1,14 +1,13 @@
 """
 Based on https://github.com/pybind/pybind11_benchmark/blob/master/setup.py
 """
-
+import signal
+import time
 import sys
-import sysconfig
 import os
 import tempfile
 import setuptools
 import subprocess
-import platform
 from setuptools import setup, Extension
 from setuptools.command.build_ext import build_ext
 
@@ -43,6 +42,49 @@ class get_numpy_include(object):
         return numpy.get_include()
 
 
+def compiler_check():
+    """
+    A very hacky method by which we understand what compiler was used to 
+    compile the user's Python.
+
+    This is necessary because setuptools will use the compiler that compiled
+    python -- even if the user specifies another one! -- for some of the
+    compilation process
+    """
+    def convert_bin_to_bytestring(bin_file):
+        MAX_BYTES = 45
+        str_ = b''
+        tmp = bin_file.read(MAX_BYTES) # can only read 45 bytes at a time
+        while tmp:
+            str_ += tmp
+            tmp = bin_file.read(MAX_BYTES)
+        return str_
+
+    with open('tmp_output.txt', 'wb') as fout, \
+        open('tmp_error.txt', 'wb') as ferr:
+        # Spawn a python3 process, then kill it
+        process = subprocess.Popen(['python3'], stdout=fout, stderr=ferr)
+        time.sleep(1) # Need to wait for python3 to produce output
+        process.send_signal(signal.SIGKILL)
+
+    with open('tmp_output.txt', 'rb') as fout, open('tmp_error.txt', 'rb') as ferr:
+        error = convert_bin_to_bytestring(ferr)
+        if b'Apple' in error:
+            raise Exception("Error: you're using a version of python compiled with Apple Clang, \
+                which does not support OpenMP. Please use Anaconda Python or download LLVM Clang \
+                and build python.")
+        elif b'Clang' in error:
+            compiler = 'clang'
+        elif b'comamnd not found' in error:
+            raise Exception("Please install python3")
+        else:
+            compiler = 'gcc'
+
+    os.remove('tmp_output.txt')
+    os.remove('tmp_error.txt')
+    return compiler
+
+
 def has_flag(compiler, flagname):
     """
     Return a boolean indicating whether a flag name is supported on
@@ -68,19 +110,18 @@ def cpp_flag(compiler):
     Return the -std=c++[11/14/17] compiler flag.
     The newer version is prefered over c++11 (when it is available).
     """
-
-    if sys.platform == "darwin":
-        # Assuming that on OSX, building with clang
+    compiler_name = compiler_check()
+    if compiler_name == 'clang':
         flags = ["-std=c++17", "-std=c++14", "-std=c++11"]
     else:
-        # Assuming that on linux, building on a manylinux image (old) with gcc
+        # Assume gcc
         flags = ["-std=c++1y"]
 
     for flag in flags:
         if has_flag(compiler, flag):
             return flag
 
-    raise RuntimeError("Unsupported compiler -- at least C++11 support " "is needed!")
+    raise RuntimeError("Unsupported compiler -- at least C++11 support is needed!")
 
 
 def check_brew_package(pkg_name):
@@ -212,7 +253,7 @@ class BuildExt(build_ext):
 
     if sys.platform == "darwin":
         install_check_mac()
-
+        assert(compiler_check() == 'clang', "Need to install LLVM clang!")
         darwin_opts = ["-stdlib=libc++", "-mmacosx-version-min=10.7", "-O3"]
         c_opts["unix"] += darwin_opts
         l_opts["unix"] += darwin_opts
@@ -236,12 +277,15 @@ class BuildExt(build_ext):
         opts.append("-fopenmp")
 
         # TODO: Change OMP library library name based on gcc vs clang instead of based on OS
+        compiler_name = compiler_check()
         if sys.platform == "darwin":
-            # We assume that if the user is on OSX, then they are building with clang (required above)
+            assert(compiler_name == 'clang', "Need to install LLVM clang!")
             link_opts.append('-lomp')
         else:
-            # We assume that if the user is on linux, then they are building with gcc
-            link_opts.append("-lomp")
+            if compiler_name == 'clang':
+                link_opts.append("-lomp")
+            else:
+                link_opts.append("-lgomp")
 
         if ct == "unix":
             if has_flag(self.compiler, "-fvisibility=hidden"):
@@ -257,7 +301,6 @@ class BuildExt(build_ext):
         build_ext.build_extensions(self)
 
 
-# TODO: Change OMP library library name based on gcc vs clang instead of based on OS
 if sys.platform == "linux" or sys.platform == "linux2":
     include_dirs = [
         get_pybind_include(),
@@ -267,8 +310,7 @@ if sys.platform == "linux" or sys.platform == "linux2":
         "headers/carma/include",
         "/usr/local/include",
     ]
-    # We assume that if the user is on linux, then they are building with gcc
-    libraries = ["armadillo", "omp"]
+    
 else:  # OSX
     include_dirs = [
         get_pybind_include(),
@@ -279,8 +321,12 @@ else:  # OSX
         "headers/carma/include/carma/carma",
         "/usr/local/include",
     ]
-    # We assume that if the user is on OSX, then they are building with clang (required above)
+
+compiler_name = compiler_check()
+if compiler_name == "clang":
     libraries = ["armadillo", "omp"]
+else: # gcc
+    libraries = ["armadillo", "gomp"]
 
 ext_modules = [
     Extension(
